% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aegrade.R
\docType{data}
\name{aegrade}
\alias{aegrade}
\title{aegrade example.}
\format{
A data frame with 20 observations on the following 6 variables:
\describe{
  \item{Dose}{Predictor}
  \item{AEgrade}{response}
}
}
\usage{
aegrade
}
\description{
A clinical trial with Dose and Adverse Event Grade output
}
\keyword{datasets}
