% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bord.R
\name{bord}
\alias{bord}
\title{ordCRM Function}
\usage{
bord(
  formula,
  data = NULL,
  family = "multinomial",
  nlev = NULL,
  nburnin = 1000,
  nmc = 5000,
  nthin = 1,
  nchain = 1,
  prior.mean = 0,
  prior.var = 1,
  alpha.min = -1000,
  alpha.max = 1000,
  newx = NULL,
  ci = 0.95,
  Pcutoffs = NULL,
  Pcutoffs3 = NULL,
  Pcutoffs4 = NULL,
  DLT.cutoff = 3,
  seed = 1
)
}
\arguments{
\item{formula}{as in \code{\link{lm}}}

\item{data}{as in \code{\link{lm}}}

\item{family}{as in \code{\link{glm}}. Only "multinomial" is supported}

\item{nlev}{number of categories in the toxicity profile. If NULL this will be counted
from the data}

\item{nburnin}{number of burnin samples}

\item{nmc}{number of markov chain monte carlo samples after burnin}

\item{nthin}{thinning interval, default is 1 (no thinning)}

\item{nchain}{number of markov chains}

\item{prior.mean}{prior mean of the regression parameter}

\item{prior.var}{prior variance of the regression parameter.}

\item{alpha.min}{the lowest point of the intercept parameter space}

\item{alpha.max}{the highest point of the intercept parameter space}

\item{newx}{new covariates for prediction. If NULL the observed data will be used}

\item{ci}{significance level to calculate Bayesian credible interval. Default is 0.95}

\item{Pcutoffs}{target toxicity interval -- a vector containg two probability elements -- first 
is the boundary point between underdose and target toxicity region, second is the boundary 
between target toxicity and overdose region}

\item{Pcutoffs3}{same as above for grade 3}

\item{Pcutoffs4}{same as above for grade 4}

\item{DLT.cutoff}{adverse event grade cut off for DLT, must be 3 or 4.
If 3 then DLT >= grade 3, if 4 then DLT >= grade 4}

\item{seed}{Seed for R for reproducibility}
}
\value{
\item{beta.post}{Posterior mean of \eqn{\beta}}
\item{beta.sd}{Posterior standard deviation of \eqn{\beta}}
\item{beta.median}{Posterior meadian of \eqn{\beta}}
\item{beta.summary}{Posterior summary of \eqn{\beta}}
\item{left.ci}{Lower point of the Bayesian credible interval for \eqn{\beta}}
\item{right.ci}{Upper point of the Bayesian credible interval for \eqn{\beta}}
\item{beta.samples}{Posterior samples of \eqn{\beta}}
\item{pTox.post}{Posterior predictive summaries of the dose levels}
\item{pCat.post}{Posterior predictive probabilities of doses}
\item{predTox.post}{Posterior predictive toxicity probilities of the future doses}
\item{pred.dose}{Doses for compound at which the prediction is computed. This is 
same as newx}
\item{alpha.post}{Posterior means of the intercepts}
\item{grade.prob.post}{Predictive probabilities of doses for each adverse event grade}
}
\description{
This function is a code to the continuous assesment method (CRM) with ordinal 
endpoints in the oncology participants data. The important assumption is proportional odds.
}
\details{
For single-agent data, the base model, is a standard cumulative logistic 
proportional odds model in log-dose. We place the Normal(prior.mean, prior.var) distribution on the 
regression paramter \eqn{\beta}
}
\examples{
\dontrun{
data(crs.pk)

y <- as.factor(crs.pk$CRS)
x <- as.vector(crs.pk$Cmax)

fit <- bord(formula   = as.factor(CRS) ~ Cmax, 
           data       = crs.pk, 
           nburnin    = 1000, 
           nmc        = 10000, 
           prior.mean = 0,
           prior.var  = 100, 
           Pcutoffs   = c(0.16, 0.33),
           Pcutoffs3  = c(0.10, 0.25),
           Pcutoffs4  = c(0.05, 0.10),
           DLT.cutoff = 3,
           newx       = 3, 
           nlev       = 5,
           alpha.min  = -1000,
           alpha.max  = 1000
           )
              
 fit$beta.post
 fit$alpha.post
}



}
\references{
Albert, James A. and Chib, Siddhartha. (1993) Bayesian Analysis of Binary
and Polychotomous Response Data. Journal of the American Statistical Association,
88(422), 669-679.

Sorensen, D. A., Anderson, S., Gianola, D., and Korsgaard, I. (1995) Bayesian 
inference in thereshold models using Gibbs sampling. Genetics Selection Evolution,
27(3), 229-249.

Polson, Nicholas G., Scott, James G., and Windle, Jesse. (2013) Bayesian Inference 
for Logistic Models Using Polya-Gamma Latent Variables, Journal of the American
Statistical Association, 108:504, 1339-1349.

Montesinos-Lopez, O. A., Montesinos-Lopez, A., Crossa, J., Burgueno, J., 
& Eskridge, K. (2015). Genomic-enabled prediction of ordinal data with Bayesian 
logistic ordinal regression. G3: Genes, Genomes, Genetics, 5(10), 2113-2126.

Maity, A. K. (2023). Bayesian Dose Finding Model using Ordinal Endpoints.
}
\seealso{
\code{\link{bordCRM}}
}
