% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bordCRM.R
\name{bordCRM}
\alias{bordCRM}
\title{bordCRM Function}
\usage{
bordCRM(
  Data = NULL,
  Agent = NULL,
  Doses = NULL,
  DoseRef = NULL,
  nlev = NULL,
  prior.mean = 0,
  prior.var = 1,
  Prior.Only = FALSE,
  Pcutoffs = c(0.16, 0.33),
  Pcutoffs3 = c(0.1, 0.25),
  Pcutoffs4 = c(0.05, 0.1),
  DLT.cutoff = 3,
  Outfile = "out_ordinal",
  Plot = TRUE,
  nburnin = 1000,
  nmc = 5000,
  nthin = 1,
  nchain = 1,
  nsample = 1000,
  int.crit = c(1, 1, 0.25),
  alpha.min = -1000,
  alpha.max = 1000,
  DIC = NULL,
  seed = 1452671
)
}
\arguments{
\item{Data}{Dose DLT data -- a data frame with the first vector as the dose limiting toxicity 
response which is ordinal and the second vector is the dose}

\item{Agent}{Compound name}

\item{Doses}{Doses for compound at which the prediction is required}

\item{DoseRef}{Dose reference for compound}

\item{nlev}{number of categories in the toxicity profile. If NULL this will be counted
from the data}

\item{prior.mean}{prior mean of the regression parameter}

\item{prior.var}{prior variance of the regression parameter.}

\item{Prior.Only}{either TRUE or FALSE. If TRUE, Data must be equal to NULL meaning that there 
is no data and posterior summaries are basically the summaries from the prior distribution. 
If FALSE, then Data must be provided and posterior will be computed.}

\item{Pcutoffs}{target toxicity interval -- a vector containg two probability elements -- first 
is the boundary point between underdose and target toxicity region, second is the boundary 
between target toxicity and overdose region}

\item{Pcutoffs3}{same as above for grade 3}

\item{Pcutoffs4}{same as above for grade 4}

\item{DLT.cutoff}{adverse event grade cut off for DLT, must be 3 or 4.
If 3 then DLT >= grade 3, if 4 then DLT >= grade 4}

\item{Outfile}{output file}

\item{Plot}{Indicator if plot will be generated}

\item{nburnin}{number of burnin samples}

\item{nmc}{number of markov chain monte carlo samples after burnin}

\item{nthin}{thinning interval, default is 1 (no thinning)}

\item{nchain}{number of markov chains}

\item{nsample}{number of samples from prior distribution. This is only required if 
Prior.Only = TRUE}

\item{int.crit}{Critical value for interval to change color}

\item{alpha.min}{the lowest point of the intercept parameter space}

\item{alpha.max}{the highest point of the intercept parameter space}

\item{DIC}{DIC to save}

\item{seed}{Seed for R for reproducibility}
}
\value{
\item{Data}{The dataset}
\item{Agent}{The agent's name}
\item{prior.mean}{Prior mean used for the analysis}
\item{prior.var}{Prior variance used for the analysis}
\item{beta}{Posterior summary of \eqn{\beta}}
\item{grade.prob}{Predictive probabilities of doses for each adverse event grade}
\item{P}{Posterior predictive summaries of the dose levels}
\item{Pcat}{Posterior predictive probabilities of doses}
\item{pred.Tox}{Posterior predictive toxicity probilities of the future doses}
\item{Doses}{Doses for compound at which the prediction is computed}
\item{nburnin}{number of burnin samples}
\item{nmc}{number of markov chain monte carlo samples after burnin}
\item{seed}{Seed for R used in the computation}
In addition, a text output containing all of the above and two plots -- posterior DLT rates and 
interval probabilities of DLT's, are produced.
}
\description{
This function is a code to the continuous assesment method (CRM) with ordinal 
endpoints in the oncology participants data. The important assumption is proportional odds.
}
\details{
For single-agent data, the base model, is a standard cumulative logistic 
proportional odds model in log-dose. We place the Normal(prior.mean, prior.var) distribution on the 
regression paramter \eqn{\beta}
}
\examples{
\dontrun{
data(aegrade)

y        <- as.factor(aegrade$AEgrade)
x        <- log(as.vector(aegrade$Dose))
DoseAdm  <- x  
Toxicity <- y
Data     <- data.frame(Toxicity, DoseAdm)
Doses    <- sort(unique(DoseAdm))  # we want to find the toxicity probability at this dose


fit <- bordCRM(Data       = Data,
               Agent      = "Example",
               Doses      = Doses,
               DoseRef    = NULL,
               prior.mean = 0,
               prior.var  = 1,
               Prior.Only = FALSE,
               Pcutoffs   = c(0.16, 0.33),
               Pcutoffs3  = c(0.10, 0.25),
               Pcutoffs4  = c(0.05, 0.10),
               DLT.cutoff = 3,
               Outfile    = "out_ordinal",
               Plot       = TRUE,
               nburnin    = 50,
               nmc        = 100,
               nlev       = 5,
               alpha.min  = -1000,
               alpha.max  = 1000,
               seed       = 1452671
               )
               
fit$Pcat 
fit$Pcat3
fit$Pcat4
fit$grade.prob
}



}
\references{
Meter, E. M. V., Garrett-Mayer, E., & Bandyopadhyay, D. (2010).
Proportional odds model for dose-finding clinical trial designs with ordinal
toxicity grading. Statistics in Medicine, 30(17), 2070-2080.

Darssan, D., Thompson, M. H., & Pettitt, A. N. (2014). 
Incorporating adverse event relatedness into dose-finding clinical trial designs.
Statistics in Medicine, 33(7), 1146-1161.

Maity, A. K. (2023). Bayesian Dose Finding Model using Ordinal Endpoints.
}
