% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs_pk.R
\docType{data}
\name{crs.pk}
\alias{crs.pk}
\title{crs.pk example.}
\format{
A data frame with 20 observations on the following 7 variables:
\describe{
  \item{Dose}{Predictor}
  \item{TREATXT}{Tretements}
  \item{SUBJID}{Subject ID}
  \item{Cmax}{PK variable}
  \item{AUClast}{PK variable}
  \item{Cavg}{PK variable}
  \item{CRS}{response}
}
}
\source{
Basu, Cynthia
}
\usage{
crs.pk
}
\description{
A clinical trial with PK and CRS output
}
\keyword{datasets}
